+++ backend/ws/client_event_handlers/ClientWantsToTurnOffMotionAlarm.cs
@@ -30,7 +30,7 @@ public class ClientWantsToTurnOffMotionAlarm : BaseEventHandler<ClientWantsToTur
     public override async Task Handle(ClientWantsToTurnOffAlarmsWindowDoorDto dto, IWebSocketConnection socket)
     {
         HistoryModel loggedEvent = _historyService.CreateHistory(dto.historyModel);
-        _unitService.SetAllWindowDoorStatus((int)Status.Disarmed);
+        _unitService.SetAllWindowDoorStatus(Status.Disarmed);
         await _mqttPublishService.AlarmTurnOffMotionPublish();
         var turnOffAlarm = new ResponseDto()
         {
